pyproject.toml
[project]
name = "face-matching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy==1.26.4",
  "opencv-python==4.11.0.86",
  "pillow==12.0.0",
  "scikit-learn==1.7.2",
  "tqdm==4.67.1",
  "mediapipe==0.10.21",
  "pymilvus==2.5.17",
  "fastapi==0.121.3",
  "uvicorn[standard]==0.38.0",
  "python-multipart==0.0.20",
  "pydantic==2.12.4",
  "sqlalchemy==2.0.44",
  "alembic==1.17.2",
  "streamlit==1.51.0",
]
